Clean up AL LHS source term contribution
* fix parallel sum for LHS field; it was missing. Therefore,
LHS field was most likey zero for parallel runs.

* clean up warnings from sloppy code, i.e., declared not used, etc

+++ src/NaluParsing.C
@@ -437,7 +437,7 @@ namespace YAML {
       } 
       
       yk.massFraction_.resize(node.size());
-      int ykSize = node.size() ;
+      size_t ykSize = node.size() ;
       for ( size_t k = 0; k < ykSize; ++k ) {
 	yk.massFraction_[k] = node[k].as<double>() ;
       }
